requirements.txt
asgiref @ file:///C:/Users/BUILDE~1/AppData/Local/Temp/abs_cchykh5sgr/croots/recipe/asgiref_1661287542702/work
attrs==22.2.0
beautifulsoup4==4.11.2
certifi @ file:///C:/b/abs_ac29jvt43w/croot/certifi_1665076682579/work/certifi
colorama==0.4.6
Django @ file:///C:/Users/BUILDE~1/AppData/Local/Temp/abs_eeor5o9jj8/croots/recipe/django_1661292534510/work
django-bootstrap4==22.3
django-crispy-forms==2.0
django-dynamic-fixture==3.1.2
django-jquery==3.1.0
django-simple-history==3.2.0
django-tenants==3.4.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.1.0
factory-boy==3.2.1
Faker==16.6.0
iniconfig==2.0.0
ordered-set==4.1.0
packaging==23.0
Pillow==9.3.0
pluggy==1.0.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
PyJWT==2.6.0
pytest==7.2.1
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2022.6
six==1.16.0
soupsieve==2.4
sqlparse @ file:///C:/b/abs_83jn6owfpc/croot/sqlparse_1668793499314/work
tomli==2.0.1
tzdata @ file:///tmp/build/80754af9/python-tzdata_1611693927741/work
wincertstore==0.2
